(manual) merge with cm branch
git-svn-id: https://pet.opendfki.de/repos/pet/branches/hmuelner@736 4200e16c-5112-0410-ac55-d7fb557a720a

+++ common/types.h
@@ -329,7 +329,7 @@ inline bool is_string_instance(type_t a) {
 /** check the validity of the attribute \a attr */
 inline bool is_attr(attr_t attr) {
   assert(attr >= 0);
-  return (attr <= attrname.size());
+  return (attr <= static_cast<int>(attrname.size()));
 }
 
 #endif
